[geometry] Move mesh_intersection code from .h to .cc.
